+++ src/main/java/br/com/gps/mecanica/services/VeiculoService.java
@@ -31,69 +31,73 @@ public class VeiculoService {
         }
 
         veiculo.setPlaca(placa);
-        veiculo.setModelo(Utils.formatarModeloVeiculo(veiculo.getModelo()));
-        veiculo.setMarca(Utils.formatarModeloVeiculo(veiculo.getMarca()));
+        veiculo.setModelo(Utils.formatarMarcaModeloVeiculo(veiculo.getModelo()));
+        veiculo.setMarca(Utils.formatarMarcaModeloVeiculo(veiculo.getMarca()));
 
         return veiculoRepository.save(veiculo);
-    }
-
-    public void delete(UUID id) {
-        try {
-            veiculoRepository.deleteById(id);
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
     }
 
     public List<VeiculoModel> get() {
         return veiculoRepository.findAll();
     }
 
-    public VeiculoModel get(UUID id) {
+    public VeiculoModel get(UUID id) throws Exception {
+        if (veiculoRepository.findById(id).isEmpty()) {
+            throw new Exception("Veículo não encontrado");
+        }
+
         return veiculoRepository.findById(id).get();
     }
 
-    public VeiculoModel getByPlaca(String placa) {
-        try {
-            return veiculoRepository.findByPlaca(placa);
-        } catch (Exception e) {
-            e.printStackTrace();
+    public VeiculoModel getByPlaca(String placa) throws Exception {
+        if (veiculoRepository.findByPlaca(placa) == null) {
+            throw new Exception("Veículo não encontrado");
         }
-        return null;
+
+        return veiculoRepository.findByPlaca(placa);
     }
 
     public List<VeiculoModel> getByMarca(String marca) {
-        try {
-            return veiculoRepository.findByMarca(marca);
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        return null;
+        return veiculoRepository.findByMarca(marca);
     }
 
     public List<VeiculoModel> getByModelo(String modelo) {
-        try {
-            return veiculoRepository.findByModelo(modelo);
-        } catch (Exception e) {
-            e.printStackTrace();
-        }
-        return null;
+        return veiculoRepository.findByModelo(modelo);
     }
 
     public VeiculoModel update(UUID id, VeiculoModel veiculo) throws Exception {
-
-        String placa = Utils.formatarPlaca(veiculo.getPlaca());
-
-        if (Utils.verificarPlaca(placa) == false) {
-            throw new Exception("Placa inválida");
+        if (veiculoRepository.findById(id).isEmpty()) {
+            throw new Exception("Veículo não encontrado");
         }
 
-        veiculo.setPlaca(placa);
-        veiculo.setModelo(Utils.formatarModeloVeiculo(veiculo.getModelo()));
-        veiculo.setMarca(Utils.formatarModeloVeiculo(veiculo.getMarca()));
+        VeiculoModel veiculoAtual = veiculoRepository.findById(id).get();
 
-        veiculo.setId(id);
+        if (veiculo.getPlaca() != null && !veiculo.getPlaca().isEmpty()) {
+            String placa = Utils.formatarPlaca(veiculo.getPlaca());
 
-        return veiculoRepository.save(veiculo);
+            if (Utils.verificarPlaca(placa) == false) {
+                throw new Exception("Placa inválida");
+            }
+
+            veiculoAtual.setPlaca(placa);
+        }
+
+        if (veiculo.getMarca() != null && !veiculo.getMarca().isEmpty()) {
+            veiculoAtual.setMarca(Utils.formatarMarcaModeloVeiculo(veiculo.getMarca()));
+        }
+
+        if (veiculo.getModelo() != null && !veiculo.getModelo().isEmpty()) {
+            veiculoAtual.setModelo(Utils.formatarMarcaModeloVeiculo(veiculo.getModelo()));
+        }
+
+        return veiculoRepository.save(veiculoAtual);
+    }
+
+    public void delete(UUID id) throws Exception {
+        if (veiculoRepository.findById(id).isEmpty()) {
+            throw new Exception("Veículo não encontrado");
+        }
+
+        veiculoRepository.deleteById(id);
     }
 }
